\encoding{utf-8}
\name{aegla_consensus}
\alias{aegla_consensus}
\docType{data}
\title{
Landmark coordinates data from consensus configuration of freshwater crabs
}
\description{
\emph{Aegla uruguayana} consensus shape coordinates.
}
\usage{data("aegla_consensus")}
\references{Diawol, V. P., Giri, F., & Collins, P. A. (2015). Shape and size variations of \emph{Aegla uruguayana} (Anomura-Aeglidae) under laboratory conditions. A geometric morphometric approach to the growth. Iheringia Série Zoologia, 105(1), 76-83.}
