\encoding{utf-8}
\name{aegla_landmarks}
\alias{aegla_landmarks}
\docType{data}
\title{
Landmark coordinates data from freshwater crabs
}
\description{
Geometric morphometric data of \emph{Aegla uruguayana} cephalothorax (21 landmarks).
}
\usage{data("aegla_landmarks")}
\references{Diawol, V. P., Giri, F., & Collins, P. A. (2015). Shape and size variations of \emph{Aegla uruguayana} (Anomura-Aeglidae) under laboratory conditions. A geometric morphometric approach to the growth. Iheringia Série Zoologia, 105(1), 76-83.}
\keyword{datasets}
