\encoding{utf-8}
\name{simulated_shape_distances}
\alias{simulated_shape_distances}
\docType{data}
\title{Procrustes distances data}
\description{
Procrustes distance between a set of simulated shapes and their ancestry over generations.
}
\usage{data("simulated_shape_distances")}
\keyword{datasets}
