\encoding{utf-8}
\name{simulated_shapes}
\alias{simulated_shapes}
\docType{data}
\title{
Simulated shapes
}
\description{
Simulated landmark data using evomorph \code{SimEvo} function on \emph{Aegla uruguayana} data.
}
\usage{data("simulated_shapes")}
\references{
Diawol, V. P., Giri, F., & Collins, P. A. (2015). Shape and size variations of \emph{Aegla uruguayana} (Anomura-Aeglidae) under laboratory conditions. A geometric morphometric approach to the growth. Iheringia Série Zoologia, 105(1), 76-83.}
\keyword{datasets}
